Adedd prototype login service, modified a todo-list a little bit

+++ todo-list/src/main/java/com/jslubowski/todolist/repository/TodoEventRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface TodoEventRepository extends CrudRepository<TodoEvent, Integer> {
+public interface TodoEventRepository extends CrudRepository<TodoEvent, Long> {
 
     @Query("select event from TodoEvent event where event.name like %:text%")
     List<TodoEvent> searchForEvents(@Param("text") String text);
